+++ graspit-reporters/packages/graspit-dashboard/src/components/utils/renderers.tsx
@@ -8,9 +8,9 @@ import WarningFilled from '@ant-design/icons/WarningFilled';
 import Icon from '@ant-design/icons';
 import LoadingOutlined from '@ant-design/icons/LoadingOutlined';
 import type { possibleBrowserNames } from 'src/types/sessionRecords';
-import Chrome from '../../public/chrome.svg';
-import Firefox from '../../public/firefox.svg';
-import Edge from '../../public/edge.svg';
+import Chrome from '../../../public/chrome.svg';
+import Firefox from '../../../public/firefox.svg';
+import Edge from '../../../public/edge.svg';
 import RelativeTo, { HumanizeDuration } from './Datetime/relativeTime';
 
 export default function RenderTimeRelativeToStart(props: {