+++ graspit-reporters/packages/graspit-dashboard/src/pages/_app.tsx
@@ -1,7 +1,7 @@
-import React, { type ReactNode } from "react";
-import type { AppProps } from "next/app";
-import withTheme from "@/components/theme";
-import "@/styles/globals.css";
+import React, { type ReactNode } from 'react';
+import type { AppProps } from 'next/app';
+import withTheme from 'src/components/theme';
+import 'src/styles/globals.css';
 
 const App = ({ Component, pageProps }: AppProps): ReactNode =>
     withTheme(<Component {...pageProps} />);