+++ graspit-reporters/packages/graspit-commons/CHANGELOG.md
@@ -1,5 +1,16 @@
 # graspit-commons
 
+## 1.1.0
+
+### Minor Changes
+
+- Improved the Import paths and types for the reporter for webdriverio
+
+### Patch Changes
+
+- Updated dependencies
+  - graspit@1.1.0
+
 ## 1.0.0
 
 ### Major Changes